requirements.txt
autopep8==1.5.5
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
click==7.1.2
configobj==5.0.6
cx-Oracle==8.1.0
decorator==4.4.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
kafka-python==2.0.2
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycodestyle==2.6.0
Pygments==2.8.0
pyzmq==22.0.3
regex==2020.11.13
six==1.15.0
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
wcwidth==0.2.5
yapf==0.30.0
